+++ locust/__init__.py
@@ -1,7 +1,9 @@
-from .core import HttpLocust, User, TaskSet, TaskSequence, task, seq_task
-from .exception import InterruptTaskSet, ResponseError, RescheduleTaskImmediately
-from .wait_time import between, constant, constant_pacing
+from .core import HttpLocust, Locust, TaskSet, task
 from .event import Events
+from .exception import InterruptTaskSet, ResponseError, RescheduleTaskImmediately
+from .sequential_taskset import SequentialTaskSet
+from .wait_time import between, constant, constant_pacing
+
 events = Events()
 
 __version__ = "0.14.5"