+++ locust/test/test_main.py
@@ -3,17 +3,17 @@ import os
 from locust import main
 from locust.argument_parser import parse_options
 from locust.main import create_environment
-from locust.core import HttpLocust, User, TaskSet
+from locust.core import HttpLocust, Locust, TaskSet
 from .testcases import LocustTestCase
 from .mock_locustfile import mock_locustfile
 
 
 class TestLoadLocustfile(LocustTestCase):
     def test_is_locust(self):
-        self.assertFalse(main.is_locust(("Locust", User)))
-        self.assertFalse(main.is_locust(("HttpLocust", HttpLocust)))
-        self.assertFalse(main.is_locust(("random_dict", {})))
-        self.assertFalse(main.is_locust(("random_list", [])))
+        self.assertFalse(main.is_locust(Locust))
+        self.assertFalse(main.is_locust(HttpLocust))
+        self.assertFalse(main.is_locust({}))
+        self.assertFalse(main.is_locust([]))
         
         class MyTaskSet(TaskSet):
             pass
@@ -21,16 +21,16 @@ class TestLoadLocustfile(LocustTestCase):
         class MyHttpLocust(HttpLocust):
             tasks = [MyTaskSet]
         
-        class MyLocust(User):
+        class MyLocust(Locust):
             tasks = [MyTaskSet]
         
-        self.assertTrue(main.is_locust(("MyHttpLocust", MyHttpLocust)))
-        self.assertTrue(main.is_locust(("MyLocust", MyLocust)))
+        self.assertTrue(main.is_locust(MyHttpLocust))
+        self.assertTrue(main.is_locust(MyLocust))
         
-        class ThriftLocust(User):
+        class ThriftLocust(Locust):
             abstract = True
         
-        self.assertFalse(main.is_locust(("ThriftLocust", ThriftLocust)))
+        self.assertFalse(main.is_locust(ThriftLocust))
     
     def test_load_locust_file_from_absolute_path(self):
         with mock_locustfile() as mocked: