+++ locust/test/test_stats.py
@@ -1,19 +1,23 @@
+import csv
 import time
 import unittest
 import re
 import os
 
 import gevent
+import mock
 import locust
-from locust.core import HttpLocust, TaskSet, task, User
+from locust.core import HttpLocust, TaskSet, task, Locust
 from locust.env import Environment
 from locust.inspectlocust import get_task_ratio_dict
+from locust.runners import LocalLocustRunner, MasterLocustRunner
 from locust.rpc.protocol import Message
 from locust.stats import CachedResponseTimes, RequestStats, StatsEntry, diff_response_time_dicts, stats_writer
 from locust.test.testcases import LocustTestCase
+from locust.wait_time import constant
 
 from .testcases import WebserverTestCase
-from .test_runners import mocked_options
+from .test_runners import mocked_options, mocked_rpc
 
 
 class TestRequestStats(unittest.TestCase):
@@ -283,7 +287,7 @@ class TestStatsPrinting(LocustTestCase):
         self.assertEqual(len(headlines), len(info[5].split()))
 
 
-class TestWriteStatCSVs(LocustTestCase):
+class TestCsvStats(LocustTestCase):
     STATS_BASE_NAME = "test"
     STATS_FILENAME = "{}_stats.csv".format(STATS_BASE_NAME)
     STATS_HISTORY_FILENAME = "{}_stats_history.csv".format(STATS_BASE_NAME)
@@ -291,23 +295,11 @@ class TestWriteStatCSVs(LocustTestCase):
 
     def setUp(self):
         super().setUp()
-        class MyUser(User):
-            setup_run_count = 0
-            task_run_count = 0
-            locust_error_count = 0
-            wait_time = locust.wait_time.constant(1)
-            @task
-            def my_task(self):
-                MyUser.task_run_count += 1
-        
-        self.environment = Environment(options=mocked_options())
-        locust.runners.locust_runner = locust.runners.LocalLocustRunner(self.environment, [MyUser])
         self.remove_file_if_exists(self.STATS_FILENAME)
         self.remove_file_if_exists(self.STATS_HISTORY_FILENAME)
         self.remove_file_if_exists(self.STATS_FAILURES_FILENAME)
 
     def tearDown(self):
-        locust.runners.locust_runner.quit()
         self.remove_file_if_exists(self.STATS_FILENAME)
         self.remove_file_if_exists(self.STATS_HISTORY_FILENAME)
         self.remove_file_if_exists(self.STATS_FAILURES_FILENAME)
@@ -316,19 +308,113 @@ class TestWriteStatCSVs(LocustTestCase):
         if os.path.exists(filename):
             os.remove(filename)
 
-    def test_write_stat_csvs(self):
-        locust.stats.write_stat_csvs(self.runner.stats, self.STATS_BASE_NAME)
+    def test_write_csv_files(self):
+        locust.stats.write_csv_files(self.environment, self.STATS_BASE_NAME)
         self.assertTrue(os.path.exists(self.STATS_FILENAME))
         self.assertTrue(os.path.exists(self.STATS_HISTORY_FILENAME))
         self.assertTrue(os.path.exists(self.STATS_FAILURES_FILENAME))
     
+    def test_write_csv_files_full_history(self):
+        locust.stats.write_csv_files(self.environment, self.STATS_BASE_NAME, full_history=True)
+        self.assertTrue(os.path.exists(self.STATS_FILENAME))
+        self.assertTrue(os.path.exists(self.STATS_HISTORY_FILENAME))
+        self.assertTrue(os.path.exists(self.STATS_FAILURES_FILENAME))
+    
+    @mock.patch("locust.stats.CSV_STATS_INTERVAL_SEC", new=0.2)
     def test_csv_stats_writer(self):
-        greenlet = gevent.spawn(stats_writer, self.runner.stats, self.STATS_BASE_NAME)
-        gevent.sleep(0.2)
+        greenlet = gevent.spawn(stats_writer, self.environment, self.STATS_BASE_NAME)
+        gevent.sleep(0.21)
         gevent.kill(greenlet)
         self.assertTrue(os.path.exists(self.STATS_FILENAME))
         self.assertTrue(os.path.exists(self.STATS_HISTORY_FILENAME))
         self.assertTrue(os.path.exists(self.STATS_FAILURES_FILENAME))
+        
+        with open(self.STATS_HISTORY_FILENAME) as f:
+            reader = csv.DictReader(f)
+            rows = [r for r in reader]
+        
+        self.assertEqual(2, len(rows))
+        self.assertEqual("Aggregated", rows[0]["Name"])
+        self.assertEqual("Aggregated", rows[1]["Name"])
+    
+    @mock.patch("locust.stats.CSV_STATS_INTERVAL_SEC", new=0.2)
+    def test_csv_stats_writer_full_history(self):
+        self.runner.stats.log_request("GET", "/", 10, content_length=666)
+        greenlet = gevent.spawn(stats_writer, self.environment, self.STATS_BASE_NAME, full_history=True)
+        gevent.sleep(0.21)
+        gevent.kill(greenlet)
+        self.assertTrue(os.path.exists(self.STATS_FILENAME))
+        self.assertTrue(os.path.exists(self.STATS_HISTORY_FILENAME))
+        self.assertTrue(os.path.exists(self.STATS_FAILURES_FILENAME))
+        
+        with open(self.STATS_HISTORY_FILENAME) as f:
+            reader = csv.DictReader(f)
+            rows = [r for r in reader]
+        
+        self.assertEqual(4, len(rows))
+        self.assertEqual("/", rows[0]["Name"])
+        self.assertEqual("Aggregated", rows[1]["Name"])
+        self.assertEqual("/", rows[2]["Name"])
+        self.assertEqual("Aggregated", rows[3]["Name"])
+    
+    def test_csv_stats_on_master_from_aggregated_stats(self):
+        # Failing test for: https://github.com/locustio/locust/issues/1315
+        with mock.patch("locust.rpc.rpc.Server", mocked_rpc()) as server:
+            master = MasterLocustRunner(self.environment, [], master_bind_host="*", master_bind_port=0)
+            server.mocked_send(Message("client_ready", None, "fake_client"))
+            
+            master.stats.get("/", "GET").log(100, 23455)
+            master.stats.get("/", "GET").log(800, 23455)
+            master.stats.get("/", "GET").log(700, 23455)
+            
+            data = {"user_count":1}
+            self.environment.events.report_to_master.fire(client_id="fake_client", data=data)
+            master.stats.clear_all()
+            
+            server.mocked_send(Message("stats", data, "fake_client"))
+            s = master.stats.get("/", "GET")
+            self.assertEqual(700, s.median_response_time)
+            
+            locust.stats.write_csv_files(self.environment, self.STATS_BASE_NAME, full_history=True)
+            self.assertTrue(os.path.exists(self.STATS_FILENAME))
+            self.assertTrue(os.path.exists(self.STATS_HISTORY_FILENAME))
+            self.assertTrue(os.path.exists(self.STATS_FAILURES_FILENAME))
+    
+    @mock.patch("locust.stats.CSV_STATS_INTERVAL_SEC", new=0.2)
+    def test_user_count_in_csv_history_stats(self):
+        start_time = int(time.time())
+        class TestUser(Locust):
+            wait_time = constant(10)
+            @task
+            def t(self):
+                self.environment.runner.stats.log_request("GET", "/", 10, 10)
+        runner = LocalLocustRunner(self.environment, [TestUser])
+        runner.start(3, 5) # spawn a user every 0.2 second
+        gevent.sleep(0.1)
+        
+        greenlet = gevent.spawn(stats_writer, self.environment, self.STATS_BASE_NAME, full_history=True)
+        gevent.sleep(0.6)
+        gevent.kill(greenlet)
+        
+        runner.stop()
+        
+        with open(self.STATS_HISTORY_FILENAME) as f:
+            reader = csv.DictReader(f)
+            rows = [r for r in reader]
+        
+        self.assertEqual(6, len(rows))
+        for i in range(3):
+            row = rows.pop(0)
+            self.assertEqual("%i" % (i + 1), row["User count"])
+            self.assertEqual("/", row["Name"])
+            self.assertEqual("%i" % (i + 1), row["# requests"])
+            self.assertGreaterEqual(int(row["Timestamp"]), start_time)
+            row = rows.pop(0)
+            self.assertEqual("%i" % (i + 1), row["User count"])
+            self.assertEqual("Aggregated", row["Name"])
+            self.assertEqual("%i" % (i + 1), row["# requests"])
+            self.assertGreaterEqual(int(row["Timestamp"]), start_time)
+
 
 class TestStatsEntryResponseTimesCache(unittest.TestCase):
     def setUp(self, *args, **kwargs):