+++ locust/web.py
@@ -23,7 +23,7 @@ from io import StringIO
 
 from . import runners
 from .runners import MasterLocustRunner
-from .stats import failures_csv, median_from_dict, requests_csv, sort_stats, stats_history_csv
+from .stats import failures_csv, median_from_dict, requests_csv, sort_stats
 from .util.cache import memoize
 from .util.rounding import proper_round
 from .util.timespan import parse_timespan
@@ -115,15 +115,6 @@ class WebUI:
         def request_stats_csv():
             response = make_response(requests_csv(self.environment.runner.stats))
             file_name = "requests_{0}.csv".format(time())
-            disposition = "attachment;filename={0}".format(file_name)
-            response.headers["Content-type"] = "text/csv"
-            response.headers["Content-disposition"] = disposition
-            return response
-        
-        @app.route("/stats/stats_history/csv")
-        def stats_history_stats_csv():
-            response = make_response(stats_history_csv(self.environment.runner.stats, False, True))
-            file_name = "stats_history_{0}.csv".format(time())
             disposition = "attachment;filename={0}".format(file_name)
             response.headers["Content-type"] = "text/csv"
             response.headers["Content-disposition"] = disposition